# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Microservicios Bancarios NTT Data
  version: 1.0.0
  description: API integral para la gestión de Clientes (Customer Service) y Finanzas (Account Service).
servers:
  - url: http://localhost:8081/api/v1
    description: Customer Service (Puerto 8081)
  - url: http://localhost:8082/api/v1
    description: Account-Movement-Report Service (Puerto 8082)
paths:
...
  /customers:
    get:
      tags: [ Customers ]
      summary: Listar todos los clientes
      responses:
        '200':
          description: OK
    post:
      tags: [ Customers ]
      summary: Crear un nuevo cliente
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: Santiago Moreta
              gender: Masculino
              age: 27
              identification: "104425848"
              address: Juan Molineros y 24 de Mayo
              phone: "099999999"
              password: "4321"
              status: true
            schema:
              $ref: '#/components/schemas/Customer'
      responses:
        '201':
          description: Cliente creado exitosamente
  /customers/{id}:
    get:
      tags: [ Customers ]
      summary: Obtener detalle de un cliente por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Datos del cliente encontrados
    put:
      tags: [ Customers ]
      summary: Actualizar cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: Santiago Andres Moreta
              gender: Masculino
              age: 28
              identification: "1004280666"
              address: 13 junio y Equinoccial ED
              phone: "096666666"
              password: "54321"
              status: true
            schema:
              $ref: '#/components/schemas/Customer'

      responses:
        '200':
          description: Cliente actualizado

    delete:
      tags: [ Customers ]
      summary: Eliminar cliente
      description: Elimina un cliente por su ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 4
      responses:
        '204':
          description: Cliente eliminado correctamente
        '404':
          description: Cliente no encontrado


  # --- ACCOUNT SERVICE ENDPOINTS ---
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Customer:
      type: object
      required:
        - identification
      properties:
        name:
          type: string
        gender:
          type: string
        age:
          type: integer
        identification:
          type: string
        address:
          type: string
        phone:
          type: string
        password:
          type: string
        status:
          type: boolean
...